+++ cmd/hidden_lake/applications/filesharer/pkg/settings/settings.go
@@ -16,9 +16,9 @@ const (
 )
 
 const (
-	CDefaultRetryNum   = 2
-	CDefaultPageOffset = 10
-	CDefaultLanguage   = ""
+	CDefaultRetryNum   = 2  // count
+	CDefaultPageOffset = 10 // count
+	CDefaultLanguage   = "" // ENG
 )
 
 const (