+++ cmd/hidden_lake/applications/messenger/pkg/app/errors.go
@@ -13,8 +13,12 @@ func (err *SAppError) Error() string {
 }
 
 var (
-	ErrRunning = &SAppError{"app running"}
-	ErrService = &SAppError{"service"}
-	ErrInitDB  = &SAppError{"init database"}
-	ErrClose   = &SAppError{"close"}
+	ErrRunning       = &SAppError{"app running"}
+	ErrService       = &SAppError{"service"}
+	ErrInitDB        = &SAppError{"init database"}
+	ErrClose         = &SAppError{"close"}
+	ErrInitConfig    = &SAppError{"init config"}
+	ErrGetPassword   = &SAppError{"get password"}
+	ErrWritePassword = &SAppError{"write password"}
+	ErrReadPassword  = &SAppError{"read password"}
 )