+++ cmd/hidden_lake/applications/messenger/pkg/app/init_database.go
@@ -5,14 +5,15 @@ import (
 	"path/filepath"
 
 	hlm_database "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/internal/database"
-	"github.com/number571/go-peer/pkg/database"
-
 	hlm_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/pkg/settings"
+	"github.com/number571/go-peer/pkg/database"
 )
 
 func (p *sApp) initDatabase() error {
 	sett := database.NewSettings(&database.SSettings{
-		FPath: filepath.Join(p.fPathTo, hlm_settings.CPathDB),
+		FPath:     filepath.Join(p.fPathTo, hlm_settings.CPathDB),
+		FPassword: p.fPasw,
+		FWorkSize: p.fConfig.GetSettings().GetWorkSizeBits(),
 	})
 	db, err := hlm_database.NewKeyValueDB(sett)
 	if err != nil {