+++ cmd/hidden_lake/applications/messenger/pkg/settings/settings.go
@@ -22,6 +22,7 @@ const (
 
 const (
 	CDefaultMessagesCapacity = (2 << 10) // count
+	CDefaultWorkSize         = 0         // bits
 	CDefaultLanguage         = ""        // ENG
 )