+++ cmd/hidden_lake/composite/pkg/app/init_app.go
@@ -35,7 +35,13 @@ import (
 	hls_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
 )
 
-func InitApp(pArgs []string, pDefaultPath, pDefaultKey string, pDefaultParallel uint64) (types.IRunner, error) {
+func InitApp(
+	pArgs []string,
+	pDefaultPath string,
+	pDefaultKey string,
+	pDefaultPasw string,
+	pDefaultParallel uint64,
+) (types.IRunner, error) {
 	inputPath := strings.TrimSuffix(flag.GetFlagValue(pArgs, "path", pDefaultPath), "/")
 
 	cfg, err := config.InitConfig(filepath.Join(inputPath, settings.CPathYML), nil)
@@ -43,7 +49,7 @@ func InitApp(pArgs []string, pDefaultPath, pDefaultKey string, pDefaultParallel 
 		return nil, utils.MergeErrors(ErrInitConfig, err)
 	}
 
-	runners, err := getRunners(cfg, pArgs, pDefaultPath, pDefaultKey, pDefaultParallel)
+	runners, err := getRunners(cfg, pArgs, pDefaultPath, pDefaultKey, pDefaultPasw, pDefaultParallel)
 	if err != nil {
 		return nil, utils.MergeErrors(ErrGetRunners, err)
 	}
@@ -56,6 +62,7 @@ func getRunners(
 	pArgs []string,
 	pDefaultPath string,
 	pDefaultKey string,
+	pDefaultPasw string,
 	pDefaultParallel uint64,
 ) ([]types.IRunner, error) {
 	var (
@@ -85,7 +92,7 @@ func getRunners(
 		case hll_settings.CServiceFullName:
 			runner, err = hll_app.InitApp(pArgs, pDefaultPath)
 		case hlm_settings.CServiceFullName:
-			runner, err = hlm_app.InitApp(pArgs, pDefaultPath)
+			runner, err = hlm_app.InitApp(pArgs, pDefaultPath, pDefaultPasw)
 		case hlf_settings.CServiceFullName:
 			runner, err = hlf_app.InitApp(pArgs, pDefaultPath)
 		case hla_common_settings.CServiceFullName: