+++ cmd/hidden_lake/service/internal/config/config.go
@@ -21,8 +21,8 @@ type SConfigSettings struct {
 
 	FMessageSizeBytes   uint64 `json:"message_size_bytes" yaml:"message_size_bytes"`
 	FKeySizeBits        uint64 `json:"key_size_bits" yaml:"key_size_bits"`
-	FWorkSizeBits       uint64 `json:"work_size_bits,omitempty" yaml:"work_size_bits,omitempty"`
 	FQueuePeriodMS      uint64 `json:"queue_period_ms" yaml:"queue_period_ms"`
+	FWorkSizeBits       uint64 `json:"work_size_bits,omitempty" yaml:"work_size_bits,omitempty"`
 	FQueueRandPeriodMS  uint64 `json:"queue_rand_period_ms,omitempty" yaml:"queue_rand_period_ms,omitempty"`
 	FLimitVoidSizeBytes uint64 `json:"limit_void_size_bytes,omitempty" yaml:"limit_void_size_bytes,omitempty"`
 	FNetworkKey         string `json:"network_key,omitempty" yaml:"network_key,omitempty"`